Buy Product Insert To Orders Table Created

+++ WWWPOS/ClientControl/ClientCart/Form_ClientPay.cs
@@ -8,6 +8,7 @@ using System.Text;
 using System.Threading.Tasks;
 using System.Windows.Forms;
 using WWWPOS.ClientControl.Products;
+using WWWPOS.MessageFolder;
 
 namespace WWWPOS.ClientControl.ClientCart
 {
@@ -46,9 +47,22 @@ namespace WWWPOS.ClientControl.ClientCart
         private void btn_Pay_Click(object sender, EventArgs e)
         {
             DataBase DB = new DataBase();
-            DB.PlaceOrder();
-            DB.OrdersTableInsertData();
-            DB.DeletesTheCartPending();
+
+            if (UserControl_ProductCart.CARTID == 0)
+            {
+                DB.OrdersInsertBuyData(UserControl_ProductCart.PRODUCTID, UserControl_ProductCart.PRODUCTImagePath, UserControl_ProductCart.PRODUCTName,
+                                       UserControl_ProductCart.PRODUCTPrice, UserControl_ProductCart.PRODUCTCategory, UserControl_ProductCart.PRODUCTQuantity,
+                                       UserControl_ProductCart.PRODUCTSize, UserControl_ProductCart.PRODUCTColor, UserControl_ProductCart.PRODUCTDescription);
+            }
+            else
+            {
+                DB.PlaceOrder();
+                DB.OrdersTableInsertData();
+                DB.DeletesTheCartPending();
+            }
+
+            SuccessMessageDialogue successMessageDialogue = new SuccessMessageDialogue("Placed Order Succesfully \n\n Go to cashier");
+            successMessageDialogue.ShowDialog();
 
             this.Hide();
             Form_ClientLandingPage F_ClientLandingPage = new Form_ClientLandingPage();
